chore(api): move open-api-doc/pole-emploi tests in shared
